Stop injecting BufferAllocator
