+++ embulk-ruby/lib/embulk/page_builder.rb
@@ -38,7 +38,7 @@ module Embulk
         # TODO get task as an argument
         task = Java::SPI::Exec.newConfigSource.load_config(Java::DynamicPageBuilder::BuilderTask.java_class)
         @page_builder = Java::DynamicPageBuilder.createWithTimestampMetadataFromBuilderTask(
-            task, Java::Injected::BufferAllocator, schema.to_java, java_page_output)
+            task, Java::BufferAllocator, schema.to_java, java_page_output)
       else
         @page_builder = java_dynamic_page_builder
       end
